Fixed an issue with an indefinite loop of command
The issue with indefinite looping through inquiring for command data was made by the way I handled exceptions: there could happen a situation when the new command would not be read, so the "com" value would have stayed the same. It led to the permanent fall into the responsible if statement (i.e. if (com == "add_dpt") {... )